// apps/My/constants/automationContent.enum.ts
export enum AutomationContentTypesEnum {
  AUDIO = "audio",
  BUTTON_TEMPLATE = "button_template",
  IMAGE = "image",
  INSTAGRAM_POST = "instagram_post",
  PRODUCT = "product",
  TEXT = "text",
  VIDEO = "video",
}

export enum AutomationContentModeEnum {
  AUTOMATION = "automation",
  REMINDER = "reminder",
}