// apps/My/types/file.ts
export namespace FileNamespace {

    export interface File {
    mimeType:    string;
    name:        string;
    url:         string;
    tubmnailUrl: string;
    size:        number;
    key:         string;
    id:          number;
    createDate:  Date;
    updateDate:  Date;
}


}